convert from mV to Gauss

+++ software/src/inc/my_adc.h
@@ -17,7 +17,7 @@ typedef enum{
 
 void adc_init(void);
 void adc_manager(void);
-uint8_t adc_read_mag(mag_idx_t mag_idx);
+uint16_t adc_read_mag(mag_idx_t mag_idx);
 
 
 #endif /* SRC_INC_MY_ADC_H_ */
